Fix: Category 삭제 기능 이미지 처리 수정

+++ src/main/java/site/dittotrip/ditto_trip/category/service/CategoryService.java
@@ -200,7 +200,9 @@ public class CategoryService {
     @Transactional(readOnly = false)
     public void removeCategory(Long categoryId) {
         Category category = categoryRepository.findById(categoryId).orElseThrow(NoSuchElementException::new);
-        s3Service.deleteFile(category.getImagePath());
+        if (category.getImagePath() != null) {
+            s3Service.deleteFile(category.getImagePath());
+        }
         categoryRepository.delete(category);
     }
